Backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
langchain
langchain-core
langchain-community
langchain-pinecone
langchain-google-genai
langchain-groq
pydantic
python-multipart
google-generativeai 